Resolve all remaining TODOs across the codebase (~65 items)
Implements breaking changes and improvements across 7 phases:

Phase 1: Merge HCTypeHTTPS/HCTypeTCPTLS into HTTP/TCP using TLS mode field
Phase 2: Rename LThreshold/UThreshold, change service.ip to VIP type,
  update BGP VIP interface to seesaw.VIP, change ipvs Weight to uint32
Phase 3: Engine core - config IP validation, IPv6-only support, VIP config
  guards, non-blocking channels, backend/destination overrides, VLAN updates,
  Backends IPC, clock-jump safe HA timer, anycast VIP filtering
Phase 4: Healthcheck - DNS TCP support, query validation, CNAME following,
  CNAME/NS/SOA comparison, configurable TLS ServerName, RADIUS skip auth
Phase 5: NCC - configurable BGP ASN, modprobe ip_vs fallback, IPv6 NAT,
  IPv6 routing, updated IPv6 socket options
Phase 6: HA/watchdog - RFC 5798 IP comparison, monotonic clock safety,
  watchdog failure tracking, process state checks, config defaults/group
Phase 7: CLI - shared node matcher, local node marking, glob filtering,
  healthcheck display; ECU - configurable TLS ServerName, untrusted stats;
  config - verification callback, DNS IP caching, configFromPeer via sync
  RPC; quagga - prefix count parsing; netlink SIZE_MAX comment

+++ ipvs/ipvs.go
@@ -107,7 +107,7 @@ func newIPVSDestination(dst *Destination) *ipvsDestination {
 		Address:        dst.Address,
 		Port:           dst.Port,
 		Flags:          dst.Flags,
-		Weight:         uint32(dst.Weight),
+		Weight:         dst.Weight,
 		UpperThreshold: dst.UpperThreshold,
 		LowerThreshold: dst.LowerThreshold,
 	}
@@ -152,7 +152,7 @@ func (ipvsDst ipvsDestination) toDestination() *Destination {
 	dst := &Destination{
 		Address:        ipvsDst.Address,
 		Port:           ipvsDst.Port,
-		Weight:         int32(ipvsDst.Weight), // TODO(jsing): uint32?
+		Weight:         ipvsDst.Weight,
 		Flags:          ipvsDst.Flags,
 		LowerThreshold: ipvsDst.LowerThreshold,
 		UpperThreshold: ipvsDst.UpperThreshold,
@@ -281,7 +281,7 @@ const (
 type Destination struct {
 	Address        net.IP
 	Port           uint16
-	Weight         int32
+	Weight         uint32
 	Flags          DestinationFlags
 	LowerThreshold uint32
 	UpperThreshold uint32
